refactor(chat): move getCredits out of lib/supabase
It's an HTTP client now (fetches /api/credits from recoup-api), not a
Supabase operation. Moves to lib/credits/getCredits.ts and drops the
.tsx extension since there is no JSX.
