+++ lib/credits/getCredits.ts
